requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=pyhd3eb1b0_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=24.3.0=pyh71513ae_0
backtesting=0.3.3=pypi_0
beautifulsoup4=4.12.3=py312haa95532_0
blas=1.0=mkl
bokeh=3.6.2=pypi_0
bottleneck=1.4.2=py312h4b0e54e_0
brotli=1.0.9=h827c3e9_9
brotli-bin=1.0.9=h827c3e9_9
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.12.31=haa95532_0
certifi=2024.12.14=py312haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py312haa95532_0
colorama=0.4.6=py312haa95532_0
comm=0.2.1=py312haa95532_0
contourpy=1.3.1=py312h214f63a_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.8.11=py312h5da7b33_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.3=h5da7b33_0
fonttools=4.51.0=py312h2bbff1b_0
freetype=2.12.1=ha860e81_0
frozendict=2.4.2=py312haa95532_0
html5lib=1.1=pyhd3eb1b0_0
icc_rt=2022.1.0=h6049295_2
idna=3.7=py312haa95532_0
importlib_resources=6.5.2=pyhd8ed1ab_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.5=py312haa95532_0
ipython=8.30.0=py312haa95532_0
jedi=0.19.2=py312haa95532_0
jinja2=3.1.5=pypi_0
jpeg=9e=h827c3e9_3
jsonschema=4.23.0=pyhd8ed1ab_1
jsonschema-specifications=2024.10.1=pyhd8ed1ab_1
jupyter_client=8.6.0=py312haa95532_0
jupyter_core=5.7.2=py312haa95532_0
kiwisolver=1.4.4=py312hd77b12b_0
lcms2=2.16=hb4a4139_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h827c3e9_9
libbrotlidec=1.0.9=h827c3e9_9
libbrotlienc=1.0.9=h827c3e9_9
libdeflate=1.17=h2bbff1b_1
libexpat=2.6.3=he0c23c2_0
libffi=3.4.4=hd77b12b_1
libiconv=1.16=h2bbff1b_3
libpng=1.6.39=h8cc25b3_0
libsodium=1.0.18=h62dcd97_0
libsqlite=3.47.2=h67fdade_0
libtiff=4.5.1=hd77b12b_0
libwebp-base=1.3.2=h3d04722_1
libxml2=2.13.5=h24da03e_0
libxslt=1.1.41=h0739af5_0
libzlib=1.2.13=h2466b09_6
lxml=5.3.0=py312h395c83e_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=3.0.2=pypi_0
matplotlib-base=3.9.2=py312he19b0ae_1
matplotlib-inline=0.1.6=py312haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.11=py312h827c3e9_0
mkl_random=1.2.8=py312h0158946_0
mplfinance=0.12.9b7=pyhd8ed1ab_0
multitasking=0.0.9=pyhd8ed1ab_0
nbformat=5.10.4=pyhd8ed1ab_1
nest-asyncio=1.6.0=py312haa95532_0
numexpr=2.10.1=py312h4cd664f_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
openjpeg=2.5.2=hae555c5_0
openssl=3.4.0=ha4e3fda_1
packaging=24.2=py312haa95532_0
pandas=2.2.2=py312h0158946_0
pandas-ta=0.3.14b=pyhd8ed1ab_0
parso=0.8.4=py312haa95532_0
peewee=3.17.7=py312h4c7c349_0
pillow=11.0.0=py312h096bfcc_1
pip=24.2=py312haa95532_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_2
platformdirs=3.10.0=py312haa95532_0
plotly=5.24.1=py312hfc267ef_0
pluggy=1.5.0=py312haa95532_0
prompt-toolkit=3.0.43=py312haa95532_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py312h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.15.1=py312haa95532_1
pyparsing=3.2.0=py312haa95532_0
pysocks=1.7.1=py312haa95532_0
pytest=7.4.4=py312haa95532_0
python=3.12.3=h2628c8c_0_cpython
python-dateutil=2.9.0post0=py312haa95532_2
python-dotenv=0.21.0=py312haa95532_0
python-fastjsonschema=2.21.1=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.12=5_cp312
pytz=2024.1=py312haa95532_0
pywin32=305=py312h2bbff1b_0
pyyaml=6.0.2=py312h827c3e9_0
pyzmq=26.2.0=py312h5da7b33_0
referencing=0.35.1=pyhd8ed1ab_1
requests=2.32.3=py312haa95532_1
rpds-py=0.10.6=py312h636fa0f_1
scipy=1.14.1=py312h9d85e7c_0
setuptools=75.1.0=py312haa95532_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.5=py312haa95532_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.8.0=h59b6b97_0
tenacity=9.0.0=py312haa95532_0
tk=8.6.14=h0416ee5_0
tornado=6.4.2=py312h827c3e9_0
tqdm=4.66.5=py312hfc267ef_0
traitlets=5.14.3=py312haa95532_0
tzdata=2024b=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_1
unicodedata2=15.1.0=py312h2bbff1b_0
urllib3=2.2.3=py312haa95532_0
vc=14.40=h2eaa2aa_1
vc14_runtime=14.42.34433=he29a5d6_23
vs2015_runtime=14.42.34433=hdffcdeb_23
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py312haa95532_2
websocket-client=1.8.0=py312haa95532_0
wheel=0.44.0=py312haa95532_0
win_inet_pton=1.1.0=py312haa95532_0
xyzservices=2024.9.0=pypi_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yfinance=0.2.51=pyhd8ed1ab_0
zeromq=4.3.5=hd77b12b_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h2466b09_6
zstd=1.5.6=h8880b57_0
